Added infrastructure to support different "programs" instead of a single main.cpp
The first argument must now be a name of a program. Each program is represented
by a function program_NAME in programs/NAME.cpp. main.cpp now only contains code
to call the right progam -- to add a new program, simply add a file in programs/,
add the .cpp file to the CMakeList.txt, and add a DECLARE_PROGRAM line to main.cpp