Fixed issue in Brownian proximity network
Previoisly, next() would update the adjacency list directly,
which meant that after calling next(), the graph already reflected
the new state. That, however, was wrong -- only step() is supposed
to change the configuration of the graph. So now, next() determines
the next event but leaves it to step() to update the adjacency list.