Greatly improved the performance of sample() for some of the distributions.
We previously relied on CDF inversion for sample(). This, however is only
necessary if t > 0 and m > 1. In the concurrenty edges case, however, only
samples of the unmodified distribution (i.e. t == 0, m == 1) are required.
In this case, inverting the CDF was unnecessarily slow. We now use the
C++ standard libraries random distributions to generate samples if t == 0
and m == 1, which speeds things up by almost a factor of 10 in some cases.