Added basic support for weighted networks
Added a base class weighted_network that defines the interface for accessing weights,
which is neighbour(node_t node, index_t neighbour, double* weight). The neighbour
at the specific index is returned and its weight is stored in <weight>.

Added weighted_erdos_reyni which is an Erdös-Reyni network in which each edge has
a weight that is drawn i.i.d from a distribution. There is a simply test in
tests/weighted_network.cpp that doesn't check much yet.

The next reaction scheme correctly deals with weights, all other algorithms ignore them.
There is a test-case in test/nextreaction.cpp that runs a simulation on a weighted
network. There is no validation of the results for now though.