Fixed simulations on weighted networks. The previous approach was wrong, have to scale the edge multiplicity instead of time
Repurposed the field previously added to network_even_t called "infitesimal_duration" for general
edge weights, and renamed it to "weight". This now holds the duration in the case of infitesimal
contacts, and otherwise the weight of the edge. This fits well since both values effectively scale
the hazard rate.